requirements.txt
pandas
requests
ta
python-telegram-bot==13.15
